requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
asgiref==3.7.2
astroid==3.1.0
asttokens==2.4.1
astunparse==1.6.3
autopep8==2.0.4
beautifulsoup4==4.12.3
black==24.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
dill==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dm-tree==0.1.8
executing==2.0.1
flatbuffers==24.3.7
gast==0.5.4
google==3.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.2
google-generativeai==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h5py==3.10.0
idna==3.6
ipython==8.22.2
isort==5.13.2
jedi==0.19.1
keras==3.0.5
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.3.2
mypy-extensions==1.0.0
namex==0.0.7
numpy==1.26.4
opt-einsum==3.3.0
packaging==23.2
parso==0.8.3
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyjslint==0.3.4
PyJWT==2.8.0
pylint==3.1.0
pytz==2024.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
rich==13.7.1
rsa==4.9
setuptools==69.1.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
tomlkit==0.12.3
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
wheel==0.42.0
wrapt==1.16.0
